Re-organize color preview components, reduce resizes for big preview
