+++ src/lib/color-preview/ColorMedium.svelte
